fix: event page calendar button position + cleanup
