+++ apps/web/src/app/events/[slug]/page.module.scss
@@ -141,14 +141,29 @@
   gap: 16px;
   align-self: stretch;
 }
- .actionBox {
-  z-index: 10;
-  width: 100%;
+
+.actionBoxWrapper {
+  z-index: 2;
+
   display: flex;
-  flex-direction: row-reverse;
-  align-items: flex-start;
-  margin-top: 80px;
-  height: 100%;
+  margin: 0 auto;
+  width: 100%;
+  justify-content: center;
+  align-self: stretch;
+
+  .actionBoxContainer {
+    width: 100%;
+    max-width: responsiveness.$content-width;
+    margin: 8px 16px 16px;
+
+    .actionBox {
+      display: flex;
+      flex-direction: row-reverse;
+      .actionButton {
+        border-radius: 8px;
+      }
+    }
+  }
 }
 
 @media only screen and (max-width: 820px) {
@@ -181,7 +196,7 @@
   .location {
     order: 2;
   }
-  f .timeBox {
+  .timeBox {
     order: 3;
   }
