+++ apps/web/src/components/atoms/calendarButton/CalendarButton.module.scss
@@ -1,8 +1,7 @@
 @use '@/colors.module.scss' as colors;
 
-.calendarButton{
+.calendarButton {
     border: 0;
     font-weight: bold !important;
     color: colors.$primary;
-
 }
